+++ src/main/java/com/example/Belog/BelogApplication.java
@@ -1,0 +1,11 @@
+package com.example.Belog;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BelogApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(BelogApplication.class, args);
+    }
+}
