+++ src/main/java/com/example/Belog/controller/HomeController.java
@@ -1,0 +1,13 @@
+package com.example.Belog.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+
+    @GetMapping("/")
+    public String home() {
+        return "home";
+    }
+}
