+++ src/main/java/com/example/Belog/domain/BoardDTO.java
@@ -1,0 +1,33 @@
+package com.example.Belog.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
+public class BoardDTO {
+    private Long boardNum;
+    private Long userId;
+    private String boardTitle;
+    // private Date writeTime;
+    private String boardContents;
+    private String hashTag;
+
+    @Builder
+    public BoardDTO(
+            Long boardNum,
+            Long userId,
+            String boardTitle,
+            // Date writeTime,
+            String boardContents,
+            String hashTag
+    ) {
+        this.boardNum = boardNum;
+        this.userId = userId;
+        this.boardTitle = boardTitle;
+        // this.writeTime = writeTime;
+        this.boardContents = boardContents;
+        this.hashTag = hashTag;
+    }
+}
