+++ src/main/java/com/example/Belog/exception/BoardException.java
@@ -1,0 +1,9 @@
+package com.example.Belog.exception;
+
+public class BoardException extends RuntimeException{
+    private static final long serialVersionUID = 1L;
+
+    public BoardException(String message) {
+        super(message);
+    }
+}
