+++ src/main/java/com/example/Belog/exception/BoardNotFoundException.java
@@ -1,0 +1,8 @@
+package com.example.Belog.exception;
+
+public class BoardNotFoundException extends BoardException{
+
+    public BoardNotFoundException(String message) {
+        super(message);
+    }
+}
