+++ src/main/java/com/example/Belog/exception/UserException.java
@@ -1,0 +1,9 @@
+package com.example.Belog.exception;
+
+public class UserException extends RuntimeException{
+    private static final long serialVersionUID = 1L;
+
+    public UserException(String message) {
+        super(message);
+    }
+}
