+++ src/main/java/com/example/Belog/mapper/BoardMapper.java
@@ -1,0 +1,24 @@
+package com.example.Belog.mapper;
+
+import com.example.Belog.domain.BoardDTO;
+import com.example.Belog.domain.Criteria;
+import org.apache.ibatis.annotations.Mapper;
+
+import java.util.List;
+
+@Mapper
+public interface BoardMapper {
+    int writeBoard(BoardDTO boardDTO);
+    
+    int deleteBoard(Long boardNum);
+
+    int updateBoard(BoardDTO boardDTO);
+
+    List<BoardDTO> findAllBoard(Criteria criteria);
+
+    int getBoardCount();
+
+    BoardDTO findBoardByBoardNum(Long boardNum);
+
+    int getBoardCountByBoardNum(Long boardNum);
+}
