--- src/main/java/com/example/velog/controller/HomeController.java
@@ -1,13 +1,0 @@
-package com.example.velog.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-
-    @GetMapping("/")
-    public String home() {
-        return "home";
-    }
-}
