--- src/main/java/com/example/velog/exception/BoardException.java
@@ -1,9 +1,0 @@
-package com.example.velog.exception;
-
-public class BoardException extends RuntimeException{
-    private static final long serialVersionUID = 1L;
-
-    public BoardException(String message) {
-        super(message);
-    }
-}
