--- src/main/java/com/example/velog/exception/BoardNotFoundException.java
@@ -1,8 +1,0 @@
-package com.example.velog.exception;
-
-public class BoardNotFoundException extends BoardException{
-
-    public BoardNotFoundException(String message) {
-        super(message);
-    }
-}
