--- src/main/java/com/example/velog/exception/UserException.java
@@ -1,9 +1,0 @@
-package com.example.velog.exception;
-
-public class UserException extends RuntimeException{
-    private static final long serialVersionUID = 1L;
-
-    public UserException(String message) {
-        super(message);
-    }
-}
