+++ src/main/resources/mybatis/mapper/BoardMapper.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.example.velog.mapper.BoardMapper">
+<mapper namespace="com.example.Belog.mapper.BoardMapper">
 
-    <insert id="writeBoard" parameterType="com.example.velog.domain.BoardDTO">
+    <insert id="writeBoard" parameterType="com.example.Belog.domain.BoardDTO">
         insert into board(userId, boardTitle, boardContents, hashTag) values (#{userId}, #{boardTitle}, #{boardContents}, #{hashTag})
     </insert>
 
@@ -11,11 +11,11 @@
         delete from board where boardNum = #{boardNum}
     </delete>
 
-    <update id="updateBoard" parameterType="com.example.velog.domain.BoardDTO">
+    <update id="updateBoard" parameterType="com.example.Belog.domain.BoardDTO">
         update board set boardContents = #{boardContents}, boardTitle = #{boardTitle}, hashTag = #{hashTag} where boardNum = #{boardNum}
     </update>
 
-    <select id="getBoardCount" parameterType="com.example.velog.domain.BoardDTO" resultType="int">
+    <select id="getBoardCount" parameterType="com.example.Belog.domain.BoardDTO" resultType="int">
         select count(*) from board;
     </select>
 
@@ -27,7 +27,7 @@
         select count(*) from board where boardNum = #{boardNum}
     </select>
 
-    <select id="findBoardByBoardNum" parameterType="Long" resultType="com.example.velog.domain.BoardDTO">
+    <select id="findBoardByBoardNum" parameterType="Long" resultType="com.example.Belog.domain.BoardDTO">
         select boardNum, userId, boardTitle, boardContents, hashTag from board where boardNum = #{boardNum};
     </select>
 </mapper>
