+++ src/main/resources/mybatis/mapper/CommentMapper.xml
@@ -1,26 +1,26 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.example.velog.mapper.CommentMapper">
+<mapper namespace="com.example.Belog.mapper.CommentMapper">
 
     <insert id="add" useGeneratedKeys="true" keyProperty="commentNum">
         INSERT INTO comment(userId, userEmail, boardNum, comment)
         VALUES (#{userId}, #{userEmail}, #{boardNum}, #{comment})
     </insert>
 
-    <update id="edit" parameterType="com.example.velog.domain.CommentDTO">
+    <update id="edit" parameterType="com.example.Belog.domain.CommentDTO">
         UPDATE comment
         SET comment = #{comment}, commentTime=sysdate
         where commentNum = #{commentNum}
     </update>
 
-    <update id="delete" parameterType="com.example.velog.domain.CommentDTO">
+    <update id="delete" parameterType="com.example.Belog.domain.CommentDTO">
         update commment
         set comment='삭제되었습니다'
         where commentNum = #{commentNum}
     </update>
 
-    <select id="findAll" resultType="com.example.velog.domain.CommentDTO">
+    <select id="findAll" resultType="com.example.Belog.domain.CommentDTO">
         SELECT boardNum, commentNum, userEmail, comment, commentTime
         FROM comment
         where boardNum = #{boardNum}
